day121/linux-log-collector/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
fastapi==0.110.2
uvicorn==0.29.0
pydantic==2.7.1
pyyaml==6.0.1
watchfiles==0.21.0
psutil==5.9.8
prometheus-client==0.20.0
structlog==24.1.0
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==5.0.0
websockets==12.0
jinja2==3.1.4
pyinotify==0.9.6
python-multipart==0.0.9
